ai-engine/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
google-generativeai>=0.3.2
pdfplumber>=0.10.3
python-docx>=1.1.0
openpyxl>=3.1.2
Pillow>=10.2.0
httpx>=0.26.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
